refactor routing
customers and products now in own routers. We might want to change basket endpoint for easier refactoring. Also database queries must be refactored
